requirements.txt
PyMuPDF
EbookLib
customtkinter
Pillow
tkinterdnd2
